requirements.txt
streamlit==1.31.0
openai==1.12.0
azure-identity==1.15.0
azure-search-documents==11.4.0
msgraph-sdk==1.2.0
python-dotenv==1.0.0
pandas==2.2.0
plotly==5.18.0
PyPDF2==3.0.1
python-docx==1.1.0
aiohttp==3.9.0
